feat: implement NetworkConnectionDatasource for socket management
- Introduced `NetworkConnectionDatasource` to manage active network connections via sockets, handling message sending and receiving.
- Refactored `NetworkService` to utilize `NetworkConnectionDatasource`, replacing the previous `NetworkConnectionManager` for improved clarity and maintainability.
- Updated methods in `NetworkService` to align with the new datasource structure, enhancing the overall network communication flow.
