+++ lib/data/services/network_service.dart
@@ -1,20 +1,19 @@
 import 'dart:io';
 import 'dart:convert';
+import 'package:jogo_da_velha/data/datasources/network_connection_datasource.dart';
 import 'package:jogo_da_velha/data/dtos/online_tic_tac_toe_game_dto.dart';
-import 'package:jogo_da_velha/domain/constants/network_message_constants.dart';
 import 'package:jogo_da_velha/domain/enums/connection_message_enum.dart';
 import 'package:jogo_da_velha/domain/enums/connection_status_enum.dart';
-import 'package:jogo_da_velha/domain/enums/config_data_key_enum.dart';
 import 'package:jogo_da_velha/domain/enums/game_message_payload_key_enum.dart';
 import 'package:jogo_da_velha/domain/enums/game_message_type_enum.dart';
-import 'package:jogo_da_velha/domain/enums/request_move_data_key_enum.dart';
 import 'package:jogo_da_velha/domain/interfaces/services/network_service_interface.dart';
 import 'package:network_info_plus/network_info_plus.dart';
 
 /// Service que executa operações de rede.
 /// Recebe/envia DTO e faz toJson/fromJson; não conhece modelo de domínio.
+/// Usa [NetworkConnectionDatasource] (camada Data) para I/O de socket.
 class NetworkService implements INetworkService {
-  NetworkConnectionManager? _connectionManager;
+  NetworkConnectionDatasource? _connectionDatasource;
 
   Function(String)? _onMessageReceived;
   Function(String)? _onConnectionStatusChanged;
@@ -34,19 +33,19 @@ class NetworkService implements INetworkService {
   @override
   set onMessageReceived(Function(String)? callback) {
     _onMessageReceived = callback;
-    _updateConnectionManagerCallbacks();
+    _updateConnectionDatasourceCallbacks();
   }
 
   @override
   set onConnectionStatusChanged(Function(String)? callback) {
     _onConnectionStatusChanged = callback;
-    _updateConnectionManagerCallbacks();
+    _updateConnectionDatasourceCallbacks();
   }
 
   @override
   set onError(Function(String)? callback) {
     _onError = callback;
-    _updateConnectionManagerCallbacks();
+    _updateConnectionDatasourceCallbacks();
   }
 
   @override
@@ -54,13 +53,13 @@ class NetworkService implements INetworkService {
     _onGameStateReceived = callback;
   }
 
-  void _updateConnectionManagerCallbacks() {
-    if (_connectionManager != null) {
-      _connectionManager!.onMessageReceived = _handleIncomingMessage;
-      _connectionManager!.onStatusChanged = (status) {
+  void _updateConnectionDatasourceCallbacks() {
+    if (_connectionDatasource != null) {
+      _connectionDatasource!.onMessageReceived = _handleIncomingMessage;
+      _connectionDatasource!.onStatusChanged = (status) {
         _onConnectionStatusChanged?.call(status.name);
       };
-      _connectionManager!.onError = (error) {
+      _connectionDatasource!.onError = (error) {
         _onError?.call(error);
       };
     }
@@ -89,7 +88,6 @@ class NetworkService implements INetworkService {
     }
   }
 
-  // Obter o IP local do dispositivo
   @override
   Future<String?> getLocalIP() async {
     try {
@@ -101,7 +99,6 @@ class NetworkService implements INetworkService {
     }
   }
 
-  /// Cria um servidor e retorna o IP local
   @override
   Future<String?> startServer({required int port}) async {
     try {
@@ -111,7 +108,7 @@ class NetworkService implements INetworkService {
         port,
       );
 
-      _connectionManager = NetworkConnectionManager(
+      _connectionDatasource = NetworkConnectionDatasource(
         serverSocket: serverSocket,
         onStatusChanged: (status) {
           _onConnectionStatusChanged?.call(status.name);
@@ -123,7 +120,7 @@ class NetworkService implements INetworkService {
       );
 
       serverSocket.listen((Socket socket) {
-        _connectionManager!.setClientSocket(socket);
+        _connectionDatasource!.setClientSocket(socket);
       });
 
       final ip = await getLocalIP();
@@ -135,7 +132,6 @@ class NetworkService implements INetworkService {
     }
   }
 
-  /// Conecta a um servidor
   @override
   Future<bool> connectToServer(String ip, {required int port}) async {
     try {
@@ -149,7 +145,7 @@ class NetworkService implements INetworkService {
       socket.setOption(SocketOption.tcpNoDelay, true);
       _onConnectionStatusChanged?.call(ConnectionStatusEnum.connected.name);
 
-      _connectionManager = NetworkConnectionManager(
+      _connectionDatasource = NetworkConnectionDatasource(
         clientSocket: socket,
         onStatusChanged: (status) {
           _onConnectionStatusChanged?.call(status.name);
@@ -161,7 +157,7 @@ class NetworkService implements INetworkService {
       );
 
       Future.delayed(const Duration(milliseconds: 100), () {
-        _connectionManager?.sendMessage('CLIENT_CONNECTED');
+        _connectionDatasource?.sendMessage('CLIENT_CONNECTED');
       });
 
       return true;
@@ -174,169 +170,32 @@ class NetworkService implements INetworkService {
 
   @override
   void disconnect() {
-    _connectionManager?.disconnect();
-    _connectionManager = null;
+    _connectionDatasource?.disconnect();
+    _connectionDatasource = null;
   }
 
   @override
   void sendReset() {
-    _connectionManager?.sendReset();
+    _connectionDatasource?.sendReset();
   }
 
   @override
   void sendNextRound() {
-    _connectionManager?.sendNextRound();
+    _connectionDatasource?.sendNextRound();
   }
 
   @override
   void sendConfig({required int maxRounds}) {
-    _connectionManager?.sendConfig(maxRounds);
+    _connectionDatasource?.sendConfig(maxRounds);
   }
 
   @override
   void sendGameState(OnlineTicTacToeGameDTO dto) {
-    _connectionManager?.sendGameState(dto.toJson());
+    _connectionDatasource?.sendGameState(dto.toJson());
   }
 
   @override
   void sendRequestMove(int row, int col) {
-    _connectionManager?.sendRequestMove(row, col);
-  }
-}
-
-/// Gerencia uma conexão de rede ativa
-/// Mantém os recursos (sockets) mas o estado é notificado via callbacks
-class NetworkConnectionManager {
-  ServerSocket? _serverSocket;
-  Socket? _clientSocket;
-  String _buffer = '';
-  Function(ConnectionStatusEnum) onStatusChanged;
-  Function(String) onMessageReceived;
-  Function(String) onError;
-
-  NetworkConnectionManager({
-    ServerSocket? serverSocket,
-    Socket? clientSocket,
-    required this.onStatusChanged,
-    required this.onMessageReceived,
-    required this.onError,
-  }) : _serverSocket = serverSocket,
-       _clientSocket = clientSocket {
-    if (_clientSocket != null) {
-      _listenToClient(_clientSocket!);
-    }
-  }
-
-  void setClientSocket(Socket socket) {
-    if (_clientSocket != null) {
-      try {
-        _clientSocket!.destroy();
-      } catch (e) {
-        // Ignora erros
-      }
-    }
-
-    socket.setOption(SocketOption.tcpNoDelay, true);
-    _clientSocket = socket;
-    onStatusChanged(ConnectionStatusEnum.connected);
-    _listenToClient(socket);
-
-    Future.microtask(() {
-      try {
-        socket.add(utf8.encode('SERVER_CONNECTED\n'));
-      } catch (e) {
-        onError('Erro ao confirmar conexão: $e');
-      }
-    });
-    onMessageReceived(NetworkMessageConstants.peerConnected);
-  }
-
-  void _listenToClient(Socket socket) {
-    socket.listen(
-      (data) {
-        _buffer += utf8.decode(data);
-        final lines = _buffer.split('\n');
-        _buffer = lines.removeLast();
-
-        for (final line in lines) {
-          if (line.trim().isNotEmpty) {
-            onMessageReceived(line.trim());
-          }
-        }
-      },
-      onError: (error) {
-        onStatusChanged(ConnectionStatusEnum.error);
-        onError('Erro na conexão: $error');
-      },
-      onDone: () {
-        onStatusChanged(ConnectionStatusEnum.disconnected);
-        _buffer = '';
-        onMessageReceived(ConnectionMessageEnum.disconnected.value);
-      },
-      cancelOnError: false,
-    );
-  }
-
-  void sendMessage(String message) {
-    if (_clientSocket != null) {
-      try {
-        _clientSocket!.add(utf8.encode('$message\n'));
-      } catch (e) {
-        onStatusChanged(ConnectionStatusEnum.error);
-        onError('Erro ao enviar mensagem: $e');
-      }
-    }
-  }
-
-  void sendGameState(Map<String, dynamic> state) {
-    final data = jsonEncode({
-      GameMessagePayloadKeyEnum.type.key: GameMessageTypeEnum.gameState.value,
-      ...state,
-    });
-    sendMessage(data);
-  }
-
-  void sendRequestMove(int row, int col) {
-    final data = jsonEncode({
-      GameMessagePayloadKeyEnum.type.key: GameMessageTypeEnum.requestMove.value,
-      RequestMoveDataKeyEnum.row.key: row,
-      RequestMoveDataKeyEnum.col.key: col,
-    });
-    sendMessage(data);
-  }
-
-  void sendReset() {
-    final data = jsonEncode({
-      GameMessagePayloadKeyEnum.type.key: GameMessageTypeEnum.reset.value,
-    });
-    sendMessage(data);
-  }
-
-  void sendNextRound() {
-    final data = jsonEncode({
-      GameMessagePayloadKeyEnum.type.key: GameMessageTypeEnum.nextRound.value,
-    });
-    sendMessage(data);
-  }
-
-  void sendConfig(int maxRounds) {
-    final data = jsonEncode({
-      GameMessagePayloadKeyEnum.type.key: GameMessageTypeEnum.config.value,
-      ConfigDataKeyEnum.maxRounds.key: maxRounds,
-    });
-    sendMessage(data);
-  }
-
-  void disconnect() {
-    try {
-      _clientSocket?.destroy();
-      _serverSocket?.close();
-    } catch (e) {
-      // Ignora erros
-    }
-    _clientSocket = null;
-    _serverSocket = null;
-    _buffer = '';
-    onStatusChanged(ConnectionStatusEnum.disconnected);
+    _connectionDatasource?.sendRequestMove(row, col);
   }
 }
